gs: Move Basic Station EUI to pkg/basicstation
